copytrade_dashboard_bot.py
import os
import time
import json
import streamlit as st
from solana.rpc.api import Client
#from solana.keypair import Keypair
from solana.account import Account as Keypair
from solana.publickey import PublicKey
from solana.transaction import Transaction
from solana.rpc.types import TxOpts
from bip_utils import Bip39SeedGenerator, Bip44, Bip44Coins
from dotenv import load_dotenv

# Load .env variables
load_dotenv()

# === Wallet Setup ===
MNEMONIC = os.getenv("PRIVATE_KEY")  # 24-word seed phrase
TARGET_WALLET = os.getenv("TARGET_WALLET") or "TARGET_WALLET_ADDRESS"

# Derive keypair from 24-word mnemonic
def get_keypair_from_mnemonic(mnemonic):
    seed_bytes = Bip39SeedGenerator(mnemonic).Generate()
    bip44 = Bip44.FromSeed(seed_bytes, Bip44Coins.SOLANA)
    priv_key = bip44.PrivateKey().Raw().ToBytes()
    return Keypair.from_secret_key(priv_key)

wallet = get_keypair_from_mnemonic(MNEMONIC)
wallet_pubkey = wallet.public_key

# === RPC and Dashboard ===
client = Client("https://api.mainnet-beta.solana.com")

st.title("📈 Solana Copy Trading Bot Dashboard")
st.write("Wallet Public Key:", str(wallet_pubkey))
st.write("Target Wallet:", TARGET_WALLET)

# === Copy-Trade Logic ===
executed_trades = set()

@st.cache_data(ttl=10)
def fetch_transactions():
    result = client.get_confirmed_signature_for_address2(PublicKey(TARGET_WALLET), limit=5)
    if not result["result"]:
        return []
    return result["result"]

def execute_trade(tx_signature):
    # Placeholder for actual swap logic, which could use Jupiter API
    # Here we simulate with a print
    st.success(f"Executed mirrored trade for {tx_signature}")

# Monitor and copy trades
transactions = fetch_transactions()
for tx in transactions:
    sig = tx["signature"]
    if sig not in executed_trades:
        execute_trade(sig)
        executed_trades.add(sig)

st.write("### Recent Trades:")
st.json(transactions)

# === Placeholder for gain logic ===
def sell_on_gain():
    # This is where you'd track token prices and implement 100% gain logic
    pass

# Run this every 10 seconds
st_autorefresh = st.empty()
while True:
    st_autorefresh.empty()
    time.sleep(5)
    transactions = fetch_transactions()
    for tx in transactions:
        sig = tx["signature"]
        if sig not in executed_trades:
            execute_trade(sig)
            executed_trades.add(sig)
    sell_on_gain()
