requirements.txt
streamlit
solana==0.31.0
solders==0.19.0
requests
python-dotenv==1.0.1
bip-utils